ai_museum/requirements.txt
asgiref==3.5.2
autopep8==1.6.0
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.5.18.1
charset-normalizer==2.0.12
colorama==0.4.5
cycler==0.11.0
Django==4.0.5
django-cors-headers==3.13.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
fonttools==4.33.3
idna==3.3
kiwisolver==1.4.3
matplotlib==3.5.2
numpy==1.23.0
opencv-python==4.6.0.66
packaging==21.3
pandas==1.4.2
Pillow==9.1.1
pycodestyle==2.8.0
PyJWT==2.4.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
requests==2.27.1
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
torch==1.12.0
torchvision==0.13.0
tqdm==4.64.0
typing_extensions==4.2.0
toml==0.10.2
tzdata==2022.1
urllib3==1.26.9
